Use symbolic links to cpp files in order to avoid referencing out-of-directory files from within the Makefile.
Previously, make would recompile every source file on each run.  This
was because it was referencing the source files from another folder,
using a relative path ../src/..  However, since it referred to the
object files as being in those folders, but actually produced the
object files in the current directory, these object files would be
recompile on every run.  This change makes symbolic links to the cpp
files (if it doesn't already exist) and then compiles them as if they
were in the build folder.